# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RRCE ERP Platform API
  description: Official Enterprise API specification for RajaRajeswari College of Engineering (RRCE) ERP System.
  version: 1.0.0
  contact:
    name: RRCE ERP Engineering Team
    email: admissions@rrce.org
servers:
  - url: https://rrce-erp.vercel.app/api
    description: Production Server (Vercel Mumbai bom1)
  - url: http://localhost:3000/api
    description: Local Development Server
paths:
...
  /auth/forgot-password:
    post:
      summary: Forgot Password Recovery
      description: Recovers institutional default password formula using USN, student name, or sequence number.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - usn
              properties:
                usn:
                  type: string
                  example: "1RR25BC007"
      responses:
        200:
          description: Recovered student identity and institutional password formula.
...
  /students/update:
    post:
      summary: Update Student Master Record
      description: Updates student master data (name, phone, DOB, quota, photoUrl) or reallocates branch. Attendance records are strictly protected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                studentId:
                  type: string
                usn:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                phone:
                  type: string
                dob:
                  type: string
                  format: date
                quota:
                  type: string
                photoUrl:
                  type: string
      responses:
        200:
          description: Student profile updated.
...